add subscription function
