+++ config.json
@@ -1,0 +1,11 @@
+{
+    "downloadVideoCover": false,
+    "downloadInterval": 1,
+    "outputDir": "./",
+    "avoidDuplicate": true,
+    "subscriptions": [
+    ],
+    "headers": {
+        "User-Agent": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/106.0.0.0 Safari/537.36"
+    }
+}
