+++ config.py
@@ -1,3 +1,15 @@
-headers = {
-    'User-Agent': "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/106.0.0.0 Safari/537.36",
-}
+import json
+
+
+def get_config(conf_path='./config.json'):
+    with open(conf_path, 'r', encoding='utf8') as f:
+        return json.load(f)
+
+
+def update_config(conf, conf_path='./config.json'):
+    with open(conf_path, 'w', encoding='utf8') as f:
+        json.dump(conf, f, indent=4, ensure_ascii=False)
+
+
+CONF = get_config()
+headers = CONF["headers"]
