+++ main.py
@@ -1,92 +1,33 @@
 # coding: utf-8
 
-import os
-import urllib.request
+import argparse
+import model_crawler
+from executor import process_subscription
 
-import cloudscraper
-import m3u8
-import requests
-from Crypto.Cipher import AES
+parser = argparse.ArgumentParser(description="Jable TV Downloader")
 
-from args import *
-from config import headers
-from cover import get_cover
-from crawler import prepareCrawl
-from delete import deleteM3u8, deleteMp4
-from merge import mergeMp4
+sub_parser = parser.add_subparsers()
 
-parser = get_parser()
-args = parser.parse_args()
+video_parser = sub_parser.add_parser("videos", help="download video by urls")
+video_parser.add_argument("urls", metavar='N', type=int, nargs='+',
+                          help="Jable TV URLs to download")
 
-if len(args.url) != 0:
-    url = args.url
-elif args.random:
-    url = av_recommand()
-else:
-    # 使用者輸入Jable網址
-    url = input('輸入jable網址:')
+models_parser = sub_parser.add_parser("subscription",
+                                      help="subscribe some topic(models or tags)/sync videos from subscriptions")
 
-# 建立番號資料夾
-urlSplit = url.split('/')
-dirName = urlSplit[-2]
-if not os.path.exists(dirName):
-    os.makedirs(dirName)
-folderPath = os.path.join(os.getcwd(), dirName)
+models_parser.add_argument("--add", type=str, default="",
+                           help="add subscription by single url, support models/tags")
+models_parser.add_argument("--get", action='store_true',
+                           help="get current subscription")
+models_parser.add_argument("--sync-videos", action='store_true',
+                           help="download all subscription related videos")
 
-# 得到 m3u8 網址
-htmlfile = cloudscraper.create_scraper(browser='firefox', delay=10).get(url)
-result = re.search("https://.+m3u8", htmlfile.text)
-m3u8url = result[0]
+models_parser.set_defaults(func=process_subscription)
 
-m3u8urlList = m3u8url.split('/')
-m3u8urlList.pop(-1)
-downloadurl = '/'.join(m3u8urlList)
+# TODO:add retry.
 
-# 儲存 m3u8 file 至資料夾
-m3u8file = os.path.join(folderPath, dirName + '.m3u8')
-urllib.request.urlretrieve(m3u8url, m3u8file)
+if __name__ == '__main__':
+    args = parser.parse_args()
 
-# 得到 m3u8 file裡的 URI和 IV
-m3u8obj = m3u8.load(m3u8file)
-m3u8uri = ''
-m3u8iv = ''
+    args.func(args)
 
-for key in m3u8obj.keys:
-    if key:
-        m3u8uri = key.uri
-        m3u8iv = key.iv
-
-# 儲存 ts網址 in tsList
-tsList = []
-for seg in m3u8obj.segments:
-    tsUrl = downloadurl + '/' + seg.uri
-    tsList.append(tsUrl)
-
-# 有加密
-if m3u8uri:
-    m3u8keyurl = downloadurl + '/' + m3u8uri  # 得到 key 的網址
-
-    # 得到 key的內容
-    response = requests.get(m3u8keyurl, headers=headers, timeout=10)
-    contentKey = response.content
-
-    vt = m3u8iv.replace("0x", "")[:16].encode()  # IV取前16位
-
-    ci = AES.new(contentKey, AES.MODE_CBC, vt)  # 建構解碼器
-else:
-    ci = ''
-
-# 刪除m3u8 file
-deleteM3u8(folderPath)
-
-# 開始爬蟲並下載mp4片段至資料夾
-prepareCrawl(ci, folderPath, tsList)
-
-# 合成mp4
-mergeMp4(folderPath, tsList)
-
-# 刪除子mp4
-deleteMp4(folderPath)
-
-# get cover
-get_cover(html_file=htmlfile, folder_path=folderPath)
