+++ utils.py
@@ -2,8 +2,7 @@ import os
 import time
 
 
-def mergeMp4(folderPath, tsList):
-    # 開始時間
+def merge_mp4(folderPath, tsList):
     start_time = time.time()
     print('開始合成影片..')
 
@@ -20,3 +19,18 @@ def mergeMp4(folderPath, tsList):
     end_time = time.time()
     print('花費 {0:.2f} 秒合成影片'.format(end_time - start_time))
     print('下載完成!')
+
+
+def deleteMp4(folderPath):
+    files = os.listdir(folderPath)
+    originFile = folderPath.split(os.path.sep)[-1] + '.mp4'
+    for file in files:
+        if file != originFile:
+            os.remove(os.path.join(folderPath, file))
+
+
+def deleteM3u8(folderPath):
+    files = os.listdir(folderPath)
+    for file in files:
+        if file.endswith('.m3u8'):
+            os.remove(os.path.join(folderPath, file))
